Initial commit for restructuring polyphase. Moved related python modules into sufolder "whatshap/polyphse". Split core algorithm from "cli/polyphase.py" into separate module "polyphase/algorithm.py.
