+++ whatshap/polyphase_solver.pxd
@@ -1,7 +1,7 @@
 # cython: language_level=3
 
 from libc.stdint cimport uint32_t, uint64_t
-from . cimport cpp
+from whatshap cimport cpp
 
 cdef class ClusterEditingSolver:
 	cdef cpp.ClusterEditingSolver *thisptr
